+++ CuttingParameterVerifier/obj/Debug/net10.0/CuttingParameterVerifier.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("CuttingParameterVerifier")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.2.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.2.0+f3eaccfba41a2d4d7548f5239e038af99dc10ee2")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.2.0+8439847efc275fd557cb792500f9f8ee55f78482")]
 [assembly: System.Reflection.AssemblyProductAttribute("CuttingParameterVerifier")]
 [assembly: System.Reflection.AssemblyTitleAttribute("CuttingParameterVerifier")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.2.0.0")]